pyproject.toml
[tool.poetry]
name = "jse-pipeline"
version = "0.1.0"
description = ""
authors = ["elroy-galbraith"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
altair = "5.1.2"
amazon-textract-caller = "0.0.29"
amazon-textract-pipeline-pagedimensions = "0.0.8"
amazon-textract-response-parser = "0.1.48"
amazon-textract-textractor = "1.3.7"
attrs = "23.1.0"
blinker = "1.6.2"
boto3 = "1.28.62"
botocore = "1.31.62"
cachetools = "5.3.1"
certifi = "2023.7.22"
cffi = "1.16.0"
chardet = "5.2.0"
charset-normalizer = "3.3.0"
click = "8.1.7"
cryptography = "41.0.4"
editdistance = "0.6.2"
gitdb = "4.0.10"
gitpython = "3.1.37"
idna = "3.4"
importlib-metadata = "6.8.0"
jinja2 = "3.1.2"
jmespath = "1.0.1"
jsonschema = "4.19.1"
jsonschema-specifications = "2023.7.1"
markdown-it-py = "3.0.0"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
mdurl = "0.1.2"
multipledispatch = "1.0.0"
natsort = "8.4.0"
numpy = "1.26.0"
packaging = "23.2"
pandas = "2.1.1"
pandas-flavor = "0.6.0"
pdf2image = "1.16.3"
pdfminer-six = "20221105"
pillow = "10.0.1"
pluginbase = "1.0.1"
protobuf = "4.24.4"
pyarrow = "13.0.0"
pycparser = "2.21"
pydeck = "0.8.1b0"
pygments = "2.16.1"
pyjanitor = "0.26.0"
pypdf = "3.16.2"
pypdf2 = "3.0.1"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
referencing = "0.30.2"
requests = "2.31.0"
rich = "13.6.0"
rpds-py = "0.10.4"
s3transfer = "0.7.0"
scipy = "1.11.3"
six = "1.16.0"
smmap = "5.0.1"
streamlit = "1.27.2"
tabulate = "0.9.0"
tenacity = "8.2.3"
textractor = "0.1.2"
toml = "0.10.2"
toolz = "0.12.0"
tornado = "6.3.3"
typing-extensions = "4.8.0"
tzdata = "2023.3"
tzlocal = "5.1"
urllib3 = "2.0.6"
validators = "0.22.0"
watchdog = "3.0.0"
xarray = "2023.9.0"
xlsxwriter = "3.0.9"
zipp = "3.17.0"
streamlit-pdf-viewer = "^0.0.14"
langchain-community = "^0.2.12"
faiss-cpu = "^1.8.0.post1"
langchain-ollama = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
